Create recipe manager interface
